Refactoring: moved GPCloudDocument to separate file
